requirements.txt
amqp==5.0.6
asgiref==3.3.4
attrs==21.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
django-cors-headers==3.7.0
django-excel-response==2.0.5
django-filter==2.4.0
django-pdfkit==0.3.1
django-templated-mail==1.1.1
django-wkhtmltopdf==3.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
flake8==3.9.2
flake8-broken-line==0.3.0
flake8-isort==4.0.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
flower==1.0.0
gunicorn==20.1.0
humanize==3.10.0
idna==2.10
inflection==0.5.1
isort==5.8.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==3.0.1
jsonschema==3.2.0
kombu==5.1.0
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
openpyxl==2.4.11
packaging==21.0
pdfkit==0.6.0
pep8-naming==0.12.1
Pillow==8.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
pyTelegramBotAPI==3.8.1
python-dotenv==0.17.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
testfixtures==6.18.0
tornado==6.1
uritemplate==3.0.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
wkhtmltopdf==0.2
